Size

+++ 0x00-hello_world/6-size.c
@@ -1,6 +1,7 @@
 #include <stdio.h>
+
 /**
- * main - Entry point
+ * main - print the sizeof many var types
  *
  * Return: Always 0 (success)
  */
@@ -12,10 +13,10 @@ int main(void)
 	char d;
 	float f;
 
-	printf("size of a char: %lu byte(s)\n", (unsigned long)sizeof(d));
-	printf("size of an int: %lu byte(s)\n", (unsigned long)sizeof(a));
-	printf("size of a long int: %lu byte(s)\n", (unsigned long)sizeof(b));
-	printf("size of a long long int: %lu byte(s)\n", (unsigned long)sizeof(c));
-	printf("size of a float: %lu byte(s)\n", (unsigned long)sizeof(f));
+	printf("size of a char: %d byte(s)\n", sizeof(d));
+	printf("size of an int: %d byte(s)\n", sizeof(a));
+	printf("size of a long int: %d byte(s)\n", sizeof(b));
+	printf("size of a long long int: %d byte(s)\n", sizeof(c));
+	printf("size of a float: %d byte(s)\n", sizeof(f));
 	return (0);
 }
